# Backendapi/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "EduSync API",
    "description": "API for EduSync Course Management System",
    "contact": {
      "name": "EduSync Support",
      "email": "support@edusync.com"
    },
    "version": "v1"
  },
  "paths": {
    "/api/Assessments": {
      "get": {
        "tags": [
          "Assessments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Assessment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Assessments"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AssessmentCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Assessment"
            }
          }
        }
      }
    },
    "/api/Assessments/{id}": {
      "get": {
        "tags": [
          "Assessments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Assessment"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Assessments"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AssessmentCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Assessments/course/{courseId}": {
      "get": {
        "tags": [
          "Assessments"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { }
            }
          }
        }
      }
    },
    "/api/Assessments/{assessmentId}/result/{studentId}": {
      "get": {
        "tags": [
          "Assessments"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assessmentId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "studentId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Results/student/{userId}/assessment/{assessmentId}": {
      "get": {
        "tags": [
          "Results"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "path",
            "name": "assessmentId",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Assessment": {
      "type": "object",
      "properties": {
        "assessmentId": {
          "format": "uuid",
          "type": "string"
        },
        "courseId": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "maxScore": {
          "format": "int32",
          "type": "integer"
        },
        "course": {
          "$ref": "#/definitions/Course"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Result"
          }
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Question"
          }
        }
      }
    },
    "AssessmentCreateDto": {
      "type": "object",
      "properties": {
        "assessmentId": {
          "format": "uuid",
          "type": "string"
        },
        "courseId": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "maxScore": {
          "format": "int32",
          "type": "integer"
        },
        "questions": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/QuestionCreateDto"
          }
        }
      }
    },
...
    "Course": {
      "type": "object",
      "properties": {
        "courseId": {
          "format": "uuid",
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "instructorId": {
          "format": "uuid",
          "type": "string"
        },
        "mediaUrl": {
          "type": "string"
        },
        "courseUrl": {
          "type": "string"
        },
        "assessments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Assessment"
          }
        },
        "instructor": {
          "$ref": "#/definitions/User"
        },
        "enrolledStudents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
...
    "Option": {
      "required": [
        "text"
      ],
      "type": "object",
      "properties": {
        "optionId": {
          "format": "uuid",
          "type": "string"
        },
        "questionId": {
          "format": "uuid",
          "type": "string"
        },
        "text": {
          "minLength": 1,
          "type": "string"
        },
        "isCorrect": {
          "type": "boolean"
        },
        "question": {
          "$ref": "#/definitions/Question"
        }
      }
    },
    "OptionCreateDto": {
      "type": "object",
      "properties": {
        "optionId": {
          "format": "uuid",
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "isCorrect": {
          "type": "boolean"
        }
      }
    },
...
    "Question": {
      "required": [
        "questionText"
      ],
      "type": "object",
      "properties": {
        "questionId": {
          "format": "uuid",
          "type": "string"
        },
        "assessmentId": {
          "format": "uuid",
          "type": "string"
        },
        "questionText": {
          "minLength": 1,
          "type": "string"
        },
        "assessment": {
          "$ref": "#/definitions/Assessment"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Option"
          }
        }
      }
    },
    "QuestionCreateDto": {
      "type": "object",
      "properties": {
        "questionId": {
          "format": "uuid",
          "type": "string"
        },
        "questionText": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OptionCreateDto"
          }
        }
      }
    },
    "Result": {
      "type": "object",
      "properties": {
        "resultId": {
          "format": "uuid",
          "type": "string"
        },
        "assessmentId": {
          "format": "uuid",
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "score": {
          "format": "int32",
          "type": "integer"
        },
        "attemptDate": {
          "format": "date-time",
          "type": "string"
        },
        "assessment": {
          "$ref": "#/definitions/Assessment"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "studentAnswers": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/StudentAnswer"
          }
        }
      }
    },
...
    "StudentAnswer": {
      "type": "object",
      "properties": {
        "answerId": {
          "format": "uuid",
          "type": "string"
        },
        "resultId": {
          "format": "uuid",
          "type": "string"
        },
        "questionId": {
          "format": "uuid",
          "type": "string"
        },
        "selectedOptionId": {
          "format": "uuid",
          "type": "string"
        },
        "assessmentId": {
          "format": "uuid",
          "type": "string"
        },
        "studentId": {
          "format": "uuid",
          "type": "string"
        },
        "result": {
          "$ref": "#/definitions/Result"
        },
        "question": {
          "$ref": "#/definitions/Question"
        },
        "selectedOption": {
          "$ref": "#/definitions/Option"
        },
        "assessment": {
          "$ref": "#/definitions/Assessment"
        },
        "student": {
          "$ref": "#/definitions/User"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "passwordHash": {
          "type": "string"
        },
        "passwordSalt": {
          "type": "string"
        },
        "courses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Course"
          }
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Result"
          }
        },
        "enrolledCourses": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Course"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}